+++ src/Task/BuiltIn/FS/CopyTask.php
@@ -28,7 +28,7 @@ class CopyTask extends AbstractFileTask
     public function getDescription()
     {
         try {
-            return sprintf('[FS] Copy %s"%s" to "%s"', $this->getFlags(), $this->getFile('from'), $this->getFile('to'));
+            return sprintf('[FS] Copy "%s" to "%s"', $this->getFile('from'), $this->getFile('to'));
         } catch (Exception $exception) {
             return '[FS] Copy [missing parameters]';
         }
@@ -38,10 +38,9 @@ class CopyTask extends AbstractFileTask
     {
         $copyFrom = $this->getFile('from');
         $copyTo = $this->getFile('to');
+        $flags = $this->options['flags'];
 
-        $flags = $this->getFlags();
-
-        $cmd = sprintf('cp %s"%s" "%s"', $flags, $copyFrom, $copyTo);
+        $cmd = sprintf('cp %s "%s" "%s"', $flags, $copyFrom, $copyTo);
 
         /** @var Process $process */
         $process = $this->runtime->runCommand($cmd);
@@ -51,11 +50,11 @@ class CopyTask extends AbstractFileTask
 
     protected function getParameters()
     {
-        return ['from', 'to'];
+        return ['from', 'to', 'flags'];
     }
 
-    protected function getDefaultFlags()
+    public function getDefaults()
     {
-        return '-p';
+        return ['flags' => '-p'];
     }
 }