+++ CHANGELOG.md
@@ -1,7 +1,7 @@
-CHANGELOG for 3.x
+CHANGELOG for 3.X
 =================
 
-* 3.0.X (2017-XX-XX)
+* 3.X (2017-XX-XX)
  * Add new Composer task, to update phar (composer/self-update)
  * [#344] Allow to flag Filesystem tasks
  * [PR#346] Add new File System task, to change file's modes (fs/chmod)